requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
python-dotenv==1.0.0